rag/app/requirements.txt
langchain==0.1.9
langchain-community==0.0.24
langsmith==0.1.10
streamlit==1.31.1
streamlit-chat==0.1.1
#chromadb==0.4.23
chromadb-client==0.4.24.dev0
pypdf==4.0.2
fastembed==0.2.2
atlassian-python-api==3.41.10
langchain-openai==0.0.8
